+++ exporter/exporterhelper/metrics_test.go
@@ -4,7 +4,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//       http://www.apache.org/licenses/LICENSE-2.0
+//	http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -29,6 +29,7 @@ import (
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/config"
+	"go.opentelemetry.io/collector/config/configtelemetry"
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/consumer/consumererror"
 	"go.opentelemetry.io/collector/internal/obsreportconfig/obsmetrics"
@@ -53,7 +54,7 @@ func TestMetricsRequest(t *testing.T) {
 	assert.EqualValues(
 		t,
 		newMetricsRequest(context.Background(), pmetric.NewMetrics(), nil),
-		mr.onError(metricsErr),
+		mr.OnError(metricsErr),
 	)
 }
 
@@ -106,7 +107,9 @@ func TestMetricsExporter_Default_ReturnError(t *testing.T) {
 }
 
 func TestMetricsExporter_WithRecordMetrics(t *testing.T) {
-	me, err := NewMetricsExporter(&fakeMetricsExporterConfig, componenttest.NewNopExporterCreateSettings(), newPushMetricsData(nil))
+	set := componenttest.NewNopExporterCreateSettings()
+	set.TelemetrySettings.MetricsLevel = configtelemetry.LevelNormal
+	me, err := NewMetricsExporter(&fakeMetricsExporterConfig, set, newPushMetricsData(nil))
 	require.NoError(t, err)
 	require.NotNil(t, me)
 
@@ -115,7 +118,9 @@ func TestMetricsExporter_WithRecordMetrics(t *testing.T) {
 
 func TestMetricsExporter_WithRecordMetrics_ReturnError(t *testing.T) {
 	want := errors.New("my_error")
-	me, err := NewMetricsExporter(&fakeMetricsExporterConfig, componenttest.NewNopExporterCreateSettings(), newPushMetricsData(want))
+	set := componenttest.NewNopExporterCreateSettings()
+	set.TelemetrySettings.MetricsLevel = configtelemetry.LevelNormal
+	me, err := NewMetricsExporter(&fakeMetricsExporterConfig, set, newPushMetricsData(want))
 	require.NoError(t, err)
 	require.NotNil(t, me)