+++ pdata/pmetric/alias.go
@@ -50,28 +50,21 @@ const (
 	MetricAggregationTemporalityCumulative = internal.MetricAggregationTemporalityCumulative
 )
 
+// MetricDataPointFlagsStruct defines how a metric aggregator reports aggregated values.
+// It describes how those values relate to the time interval over which they are aggregated.
+// Deprecated [0.58.0] Use MetricDataPointFlags instead
+type MetricDataPointFlagsStruct = MetricDataPointFlags
+
+// NewMetricDataPointFlagsStruct returns a new empty MetricDataPointFlagsStruct.
+// Deprecated [0.58.0] Use NewMetricDataPointFlags instead
+var NewMetricDataPointFlagsStruct = NewMetricDataPointFlags
+
 // MetricDataPointFlags defines how a metric aggregator reports aggregated values.
 // It describes how those values relate to the time interval over which they are aggregated.
 type MetricDataPointFlags = internal.MetricDataPointFlags
 
-const (
-	// MetricDataPointFlagsNone is the default MetricDataPointFlags.
-	MetricDataPointFlagsNone = internal.MetricDataPointFlagsNone
-)
-
-// NewMetricDataPointFlags returns a new MetricDataPointFlags combining the flags passed
-// in as parameters.
+// NewMetricDataPointFlags returns a new empty MetricDataPointFlags.
 var NewMetricDataPointFlags = internal.NewMetricDataPointFlags
-
-// MetricDataPointFlag allow users to configure DataPointFlags. This is achieved via NewMetricDataPointFlags.
-// The separation between MetricDataPointFlags and MetricDataPointFlag exists to prevent users accidentally
-// comparing the value of individual flags with MetricDataPointFlags. Instead, users must use the HasFlag method.
-type MetricDataPointFlag = internal.MetricDataPointFlag
-
-const (
-	// MetricDataPointFlagNoRecordedValue is flag for a metric aggregator which reports changes since last report time.
-	MetricDataPointFlagNoRecordedValue = internal.MetricDataPointFlagNoRecordedValue
-)
 
 // NumberDataPointValueType specifies the type of NumberDataPoint value.
 type NumberDataPointValueType = internal.NumberDataPointValueType