+++ service/collector_windows.go
@@ -95,7 +95,9 @@ func (s *windowsService) start(elog *eventlog.Log, colErrorChannel chan error) e
 	if err := s.flags.Parse(os.Args[1:]); err != nil {
 		return err
 	}
-	featuregate.GetRegistry().Apply(gatesList)
+	if err := featuregate.GetRegistry().Apply(gatesList); err != nil {
+		return err
+	}
 	var err error
 	s.col, err = newWithWindowsEventLogCore(s.settings, s.flags, elog)
 	if err != nil {
@@ -145,9 +147,9 @@ func newWithWindowsEventLogCore(set CollectorSettings, flags *flag.FlagSet, elog
 		var err error
 		cfgSet := newDefaultConfigProviderSettings(getConfigFlag(flags))
 		// Append the "overwrite properties converter" as the first converter.
-		cfgSet.MapConverters = append(
+		cfgSet.ResolverSettings.Converters = append(
 			[]confmap.Converter{overwritepropertiesconverter.New(getSetFlag(flags))},
-			cfgSet.MapConverters...)
+			cfgSet.ResolverSettings.Converters...)
 		set.ConfigProvider, err = NewConfigProvider(cfgSet)
 		if err != nil {
 			return nil, err