+++ service/internal/pipelines/pipelines_test.go
@@ -22,9 +22,6 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
-	"go.uber.org/zap/zaptest/observer"
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/component/componenttest"
@@ -348,36 +345,6 @@ func TestFailToStartAndShutdown(t *testing.T) {
 	}
 }
 
-func TestLogStabilityLevle(t *testing.T) {
-	tests := []struct {
-		level        zapcore.Level
-		expectedLogs int
-	}{
-		{
-			level:        zapcore.DebugLevel,
-			expectedLogs: 7,
-		},
-		{
-			level:        zapcore.InfoLevel,
-			expectedLogs: 3,
-		},
-	}
-
-	for _, tt := range tests {
-		observed, logs := observer.New(tt.level)
-		logger := zap.New(observed)
-		// ensure log levels are set correctly for each stability level
-		logStabilityMessage(logger, component.StabilityLevelUndefined)
-		logStabilityMessage(logger, component.StabilityLevelUnmaintained)
-		logStabilityMessage(logger, component.StabilityLevelDeprecated)
-		logStabilityMessage(logger, component.StabilityLevelInDevelopment)
-		logStabilityMessage(logger, component.StabilityLevelAlpha)
-		logStabilityMessage(logger, component.StabilityLevelBeta)
-		logStabilityMessage(logger, component.StabilityLevelStable)
-		require.Equal(t, tt.expectedLogs, logs.Len())
-	}
-}
-
 func newBadReceiverFactory() component.ReceiverFactory {
 	return component.NewReceiverFactory("bf", func() config.Receiver {
 		return &struct {
@@ -418,13 +385,13 @@ func newErrReceiverFactory() component.ReceiverFactory {
 	},
 		component.WithTracesReceiver(func(context.Context, component.ReceiverCreateSettings, config.Receiver, consumer.Traces) (component.TracesReceiver, error) {
 			return &errComponent{}, nil
-		}),
+		}, component.StabilityLevelUndefined),
 		component.WithLogsReceiver(func(context.Context, component.ReceiverCreateSettings, config.Receiver, consumer.Logs) (component.LogsReceiver, error) {
 			return &errComponent{}, nil
-		}),
+		}, component.StabilityLevelUndefined),
 		component.WithMetricsReceiver(func(context.Context, component.ReceiverCreateSettings, config.Receiver, consumer.Metrics) (component.MetricsReceiver, error) {
 			return &errComponent{}, nil
-		}),
+		}, component.StabilityLevelUndefined),
 	)
 }
 
@@ -438,13 +405,13 @@ func newErrProcessorFactory() component.ProcessorFactory {
 	},
 		component.WithTracesProcessor(func(context.Context, component.ProcessorCreateSettings, config.Processor, consumer.Traces) (component.TracesProcessor, error) {
 			return &errComponent{}, nil
-		}),
+		}, component.StabilityLevelUndefined),
 		component.WithLogsProcessor(func(context.Context, component.ProcessorCreateSettings, config.Processor, consumer.Logs) (component.LogsProcessor, error) {
 			return &errComponent{}, nil
-		}),
+		}, component.StabilityLevelUndefined),
 		component.WithMetricsProcessor(func(context.Context, component.ProcessorCreateSettings, config.Processor, consumer.Metrics) (component.MetricsProcessor, error) {
 			return &errComponent{}, nil
-		}),
+		}, component.StabilityLevelUndefined),
 	)
 }
 
@@ -458,13 +425,13 @@ func newErrExporterFactory() component.ExporterFactory {
 	},
 		component.WithTracesExporter(func(context.Context, component.ExporterCreateSettings, config.Exporter) (component.TracesExporter, error) {
 			return &errComponent{}, nil
-		}),
+		}, component.StabilityLevelUndefined),
 		component.WithLogsExporter(func(context.Context, component.ExporterCreateSettings, config.Exporter) (component.LogsExporter, error) {
 			return &errComponent{}, nil
-		}),
+		}, component.StabilityLevelUndefined),
 		component.WithMetricsExporter(func(context.Context, component.ExporterCreateSettings, config.Exporter) (component.MetricsExporter, error) {
 			return &errComponent{}, nil
-		}),
+		}, component.StabilityLevelUndefined),
 	)
 }